New FieldUtil tests
