+++ dtogen/src/test/java/io/github/manoelcampos/dtogen/util/TypeUtilTest.java
@@ -4,6 +4,7 @@ import com.karuslabs.elementary.junit.Tools;
 import com.karuslabs.elementary.junit.ToolsExtension;
 import com.karuslabs.elementary.junit.annotations.Processors;
 import io.github.manoelcampos.dtogen.DTOProcessor;
+import io.github.manoelcampos.dtogen.TestUtil;
 import io.github.manoelcampos.dtogen.samples.Class1;
 import io.github.manoelcampos.dtogen.samples.SampleClass;
 import org.junit.jupiter.api.BeforeEach;
@@ -20,7 +21,6 @@ import java.util.List;
 
 import static io.github.manoelcampos.dtogen.TestUtil.findField;
 import static io.github.manoelcampos.dtogen.util.TypeUtil.*;
-import static java.util.stream.Collectors.toMap;
 import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith(ToolsExtension.class)
@@ -36,14 +36,6 @@ class TypeUtilTest {
     void setUp() {
         Mockito.when(env.getTypeUtils()).thenReturn(Tools.types());
         this.processor = new DTOProcessor(env);
-    }
-
-    @Test
-    void testGetUpCaseFieldName() {
-        assertEquals("Name", FieldUtil.getUpCaseFieldName("name"));
-        assertEquals("Name", FieldUtil.getUpCaseFieldName("Name"));
-        assertEquals("NAME", FieldUtil.getUpCaseFieldName("nAME"));
-        assertEquals("NamE", FieldUtil.getUpCaseFieldName("namE"));
     }
 
     @Test
@@ -72,33 +64,12 @@ class TypeUtilTest {
     }
 
     @Test
-    void testIsInstanceField() {
-        final var classType = getClassTypeElement(Class1.class);
-        final var fieldsMap = classType.getEnclosedElements().stream()
-                .filter(e -> e.getKind().isField())
-                .collect(toMap(e -> e.getSimpleName().toString(), e -> (VariableElement) e));
-
-        assertFalse(FieldUtil.isInstanceField(fieldsMap.get("MAX")));
-        assertTrue(FieldUtil.isInstanceField(fieldsMap.get("id")));
-        assertTrue(FieldUtil.isInstanceField(fieldsMap.get("class2")));
-    }
-
-    @Test
     void testGetClassFields() {
         final var fieldNames = getClassFields(types, getClassTypeElement(Class1.class))
                 .map(e -> e.getSimpleName().toString())
                 .toList();
 
         assertEquals(List.of("id", "class2"), fieldNames);
-    }
-
-    @Test
-    void testIsPrimitive() {
-        final var idField = findField(elements, Class1.class, "id");
-        final var class2Field = findField(elements, Class1.class, "class2");
-
-        assertTrue(FieldUtil.isPrimitive(idField));
-        assertFalse(FieldUtil.isPrimitive(class2Field));
     }
 
     @Test
@@ -158,6 +129,6 @@ class TypeUtilTest {
     }
 
     private TypeElement getClassTypeElement(final Class<?> clazz) {
-        return elements.getTypeElement(clazz.getName());
+        return TestUtil.getClassTypeElement(elements, clazz);
     }
 }
